family-service/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
Flask-Cors==4.0.0
